Feat: Convert Time showcase pages to TS
