+++ showcase/app/templates/page-components/time.hbs
@@ -43,14 +43,14 @@
   </Shw::Flex>
 </section>
 
-<Shw::Divider {{style marginTop="6rem"}} @level="2" />
+<Shw::Divider {{style marginTop="6rem"}} @level={{2}} />
 
 <section>
   <Shw::Text::H3>Near time examples using “relative” display type</Shw::Text::H3>
 
   <Shw::Flex @gap="2rem" as |SF|>
     {{#each this.nearTimeDates as |date|}}
-      <SF.Item @label={{date}}>
+      <SF.Item @label="{{date}}">
         <Hds::Time @date={{date}} @display="relative" />
       </SF.Item>
     {{/each}}
